* add a close() function * use Py_ssize_t as type for fileSize and change default size to -1

+++ src/python/src/pypakfile.h
@@ -17,7 +17,7 @@ struct Py_PakFile {
     PyObject_HEAD
     std::ifstream *stream;
     eastwood::PakFile *pakFile;
-    uint32_t fileSize;
+    Py_ssize_t fileSize;
     file_mode mode;
 };
